+++ lib/smart_core/schema/checker/rules/optional.rb
@@ -18,8 +18,10 @@ class SmartCore::Schema::Checker::Rules::Optional < SmartCore::Schema::Checker::
   # @api private
   # @since 0.1.0
   # @version 0.3.0
+  # rubocop:disable Style/SuperArguments
   def initialize(root_reconciler, schema_key, &nested_definitions)
     super(root_reconciler, schema_key, &nested_definitions)
     @requirement = SmartCore::Schema::Checker::Rules::Requirement::Optional.new(self)
   end
+  # rubocop:enable Style/SuperArguments
 end