+++ lib/smart_core/schema/type_system/interop/operation.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+# @abstract
+# @api private
+# @since 0.12.0
+class SmartCore::Schema::TypeSystem::Interop::Operation
+  # @param type [Any]
+  # @return [void]
+  #
+  # @api private
+  # @since 0.12.0
+  def initialize(type)
+    @type = type
+  end
+
+  # @param value [Any]
+  # @return [Any]
+  #
+  # @api private
+  # @since 0.12.0
+  def call(value); end
+
+  private
+
+  # @return [Any]
+  #
+  # @api private
+  # @since 0.12.0
+  attr_reader :type
+end