+++ lib/smart_core/schema/type_system/smart_types/operation/base.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module SmartCore::Schema::TypeSystem
+  # @api private
+  # @since 0.12.0
+  class SmartTypes::Operation::Base < Interop::Operation
+  end
+end