+++ lib/smart_core/schema/type_system/smart_types/operation/valid.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module SmartCore::Schema::TypeSystem::SmartTypes::Operation
+  # @api private
+  # @since 0.12.0
+  class Valid < Base
+    # @param value [Any]
+    # @return [Boolean]
+    #
+    # @api private
+    # @since 0.12.0
+    def call(value)
+      type.valid?(value)
+    end
+  end
+end